fixing test and cleanup
